Don't specify the doctrine connection in services.yml
Specifiying the connection forces the user to override it in configuraiton
if they don't use default as the name of their connection. Symfony already
provides a configurable interface for the name of the connection as
doctrine.dbal.default_connection, and this will be used if one isn't
specified in the service. This value defaults to default if not specfied.